FirstApiPython/requirements.txt
Flask==2.0.3
azure-identity==1.7.0